+++ Request/Request.hpp
@@ -152,7 +152,7 @@ namespace ws
                 if (the_end)
                 {
                     verifyChunk(req.body);
-                    if (atoi(server_.get_body_size().c_str()) < (int)req.body.length())
+                    if (std::atoll(server_.get_body_size().c_str()) < (long long)req.body.length())
                     {
                         server_.setStatus(413);
                         return true;
